[#10] Each class a file according to its needs
