+++ lib/tryouts.rb
@@ -1,9 +1,12 @@
 # frozen_string_literal: true
 
-require 'ostruct'
 
 TRYOUTS_LIB_HOME = __dir__ unless defined?(TRYOUTS_LIB_HOME)
 
+require_relative 'tryouts/console'
+require_relative 'tryouts/section'
+require_relative 'tryouts/testbatch'
+require_relative 'tryouts/testcase'
 require_relative 'tryouts/version'
 
 class Tryouts
@@ -241,259 +244,4 @@ class Tryouts
     end
   end
 
-  class TestBatch < Array
-    class Container
-      def metaclass
-        class << self; end
-      end
-    end
-    attr_reader :path, :failed, :lines
-
-    def initialize(p, l)
-      @path = p
-      @lines = l
-      @container = Container.new.metaclass
-      @run = false
-    end
-
-    def run(before_test, &after_test)
-      return if empty?
-
-      setup
-      ret = self.select do |tc|
-        before_test.call(tc) unless before_test.nil?
-        ret = !tc.run
-        after_test.call(tc)
-        ret # select failed tests
-      end
-      @failed = ret.size
-      @run = true
-      clean
-      !failed?
-    end
-
-    def failed?
-      !@failed.nil? && @failed > 0
-    end
-
-    def setup
-      return if empty?
-
-      start = first.desc.nil? ? first.test.first : first.desc.first - 1
-      Tryouts.eval lines[0..start - 1].join, path, 0 if start > 0
-    end
-
-    def clean
-      return if empty?
-
-      last_line = last.exps.last + 1
-      return unless last_line < lines.size
-
-      Tryouts.eval lines[last_line..-1].join, path, last_line
-    end
-
-    def run?
-      @run
-    end
-  end
-
-  class TestCase
-    attr_reader :desc, :test, :exps, :failed, :passed, :skipped
-
-    def initialize(d, t, e)
-      @desc, @test, @exps, @path = d, t, e
-    end
-
-    def inspect
-      [@desc.inspect, @test.inspect, @exps.inspect].join
-    end
-
-    def to_s
-      [@desc.to_s, @test.to_s, @exps.to_s].join
-    end
-
-    def run
-      Tryouts.debug format('%s:%d', @test.path, @test.first)
-      Tryouts.debug inspect, $/
-      expectations = exps.collect do |exp, _idx|
-        exp =~ /\A\#?\s*=>\s*(.+)\Z/
-        ::Regexp.last_match(1) # this will be nil if the expectation is commented out
-      end
-
-      # Evaluate test block only if there are valid expectations
-      unless expectations.compact.empty?
-        test_value = Tryouts.eval @test.to_s, @test.path, @test.first
-        @has_run = true
-      end
-
-      @passed = []
-      @failed = []
-      @skipped = []
-      expectations.each_with_index do |exp, idx|
-        if exp.nil?
-          @skipped << '     [skipped]'
-        else
-          exp_value = Tryouts.eval(exp, @exps.path, @exps.first + idx)
-          if test_value == exp_value
-            @passed << format('     ==  %s', test_value.inspect)
-          else
-            @failed << format('     !=  %s', test_value.inspect)
-          end
-        end
-      end
-      Tryouts.debug
-      @failed.empty?
-    end
-
-    def skipped?
-      !@skipped.nil? && !@skipped.empty?
-    end
-
-    def run?
-      @has_run == true
-    end
-
-    def failed?
-      !@failed.nil? && !@failed.empty?
-    end
-
-    private
-
-    def create_proc(str, path, line)
-      eval("Proc.new {\n  #{str}\n}", binding, __FILE__, __LINE__)
-    end
-  end
-
-  class Section < Array
-    attr_accessor :path, :first, :last
-
-    def initialize(path, start = 0)
-      @path = path
-      @first = start
-      @last = start
-    end
-
-    def range
-      @first..@last
-    end
-
-    def inspect
-      range.to_a.zip(self).collect do |line|
-        "%-4d %s\n" % line
-      end.join
-    end
-
-    def to_s
-      join($/)
-    end
-  end
-
-  module Console
-    # ANSI escape sequence numbers for text attributes
-    unless defined? ATTRIBUTES
-      ATTRIBUTES = {
-        normal: 0,
-        bright: 1,
-        dim: 2,
-        underline: 4,
-        blink: 5,
-        reverse: 7,
-        hidden: 8,
-        default: 0
-      }.freeze
-    end
-
-    # ANSI escape sequence numbers for text colours
-    unless defined? COLOURS
-      COLOURS = {
-        black: 30,
-        red: 31,
-        green: 32,
-        yellow: 33,
-        blue: 34,
-        magenta: 35,
-        cyan: 36,
-        white: 37,
-        default: 39,
-        random: 30 + rand(10).to_i
-      }.freeze
-    end
-
-    # ANSI escape sequence numbers for background colours
-    unless defined? BGCOLOURS
-      BGCOLOURS = {
-        black: 40,
-        red: 41,
-        green: 42,
-        yellow: 43,
-        blue: 44,
-        magenta: 45,
-        cyan: 46,
-        white: 47,
-        default: 49,
-        random: 40 + rand(10).to_i
-      }.freeze
-    end
-
-    module InstanceMethods
-      def bright
-        Console.bright(self)
-      end
-
-      def reverse
-        Console.reverse(self)
-      end
-
-      def color(col)
-        Console.color(col, self)
-      end
-
-      def att(col)
-        Console.att(col, self)
-      end
-
-      def bgcolor(col)
-        Console.bgcolor(col, self)
-      end
-    end
-
-    def self.bright(str)
-      str = [style(ATTRIBUTES[:bright]), str, default_style].join
-      str.extend Console::InstanceMethods
-      str
-    end
-
-    def self.reverse(str)
-      str = [style(ATTRIBUTES[:reverse]), str, default_style].join
-      str.extend Console::InstanceMethods
-      str
-    end
-
-    def self.color(col, str)
-      str = [style(COLOURS[col]), str, default_style].join
-      str.extend Console::InstanceMethods
-      str
-    end
-
-    def self.att(name, str)
-      str = [style(ATTRIBUTES[name]), str, default_style].join
-      str.extend Console::InstanceMethods
-      str
-    end
-
-    def self.bgcolor(col, str)
-      str = [style(ATTRIBUTES[col]), str, default_style].join
-      str.extend Console::InstanceMethods
-      str
-    end
-
-    def self.style(*att)
-      # => \e[8;34;42m
-      "\e[%sm" % att.join(';')
-    end
-
-    def self.default_style
-      style(ATTRIBUTES[:default], ATTRIBUTES[:COLOURS], ATTRIBUTES[:BGCOLOURS])
-    end
-  end
 end
