+++ lib/tryouts/version.rb
@@ -1,13 +1,15 @@
+# frozen_string_literal: true
+
 class Tryouts
   module VERSION
     def self.to_s
-      load_config
-      [@version[:MAJOR], @version[:MINOR], @version[:PATCH]].join('.')
+      version_file
+      [@version_file[:MAJOR], @version_file[:MINOR], @version_file[:PATCH]].join('.')
     end
     alias_method :inspect, :to_s
-    def self.load_config
+    def self.version_file
       require 'yaml'
-      @version ||= YAML.load_file(File.join(TRYOUTS_LIB_HOME, '..', 'VERSION.yml'))
+      @version_file ||= YAML.load_file(File.join(TRYOUTS_LIB_HOME, '..', 'VERSION.yml'))
     end
   end
 end
